Restrict readRegister's cmd to uint8_t
It will be put into an uint8_t array anyway, so restricting the parameter to 8
bits should have no effect. Other than fixing a warning.

+++ KeyboardioScanner.cpp
@@ -87,7 +87,7 @@ byte KeyboardioScanner::setLEDSPIFrequency(byte frequency) {
 
 
 
-int KeyboardioScanner::readRegister(int cmd) {
+int KeyboardioScanner::readRegister(uint8_t cmd) {
 
     byte return_value = 0;
